мало что починиф, но нужен тупо сейф потому что я поломаф всё и кое-как починиф
